playstack:remove playstack interface
As all of functions which exist in playstack interface were moved
to playsync interface, there are no need to maintain it.

So, after moving PlayStackLayer enum to playsync interface finally,
it remove playstack interface.
