--- include/clientkit/playstack_manager_interface.hh
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef __NUGU_PLAYSTACK_MANAGER_INTERFACE_H__
-#define __NUGU_PLAYSTACK_MANAGER_INTERFACE_H__
-
-#include <map>
-#include <string>
-#include <vector>
-
-#include <base/nugu_directive.h>
-
-namespace NuguClientKit {
-
-/**
- * @file playstack_manager_interface.hh
- * @defgroup PlayStackManagerInterface PlayStackManagerInterface
- * @ingroup SDKNuguClientKit
- * @brief PlayStackManager Interface
- *
- * Interface of PlayStackManager which manage play service id in playStackControl.
- *
- * @{
- */
-
-/**
- * @brief PlayStack Layer Type
- */
-enum class PlayStackLayer {
-    None, /**< No Layer */
-    Alert, /**< Alert Layer */
-    Call, /**< Call Layer */
-    Info, /**< Info Layer */
-    Media /**< Media Layer */
-};
-
-/**
- * @}
- */
-
-} // NuguClientKit
-
-#endif /* __NUGU_PLAYSTACK_MANAGER_INTERFACE_H__ */
